Move Label Implementation and Derive Code
Move the label implementation and derive code to bevy_utils
and bevy_macro_utils crates to make it easier to reuse in other
crates.
