+++ crates/bevy_macro_utils/Cargo.toml
@@ -15,3 +15,4 @@ keywords = ["bevy"]
 [dependencies]
 cargo-manifest = "0.2.3"
 syn = "1.0"
+quote = "1.0"
