+++ crates/bevy_macro_utils/src/lib.rs
@@ -2,6 +2,7 @@ extern crate proc_macro;
 
 use cargo_manifest::{DepsSet, Manifest};
 use proc_macro::TokenStream;
+use quote::quote;
 use std::{env, path::PathBuf};
 
 pub struct BevyManifest {
@@ -58,4 +59,30 @@ fn get_path(path: &str) -> syn::Path {
 
 fn parse_str<T: syn::parse::Parse>(path: &str) -> T {
     syn::parse(path.parse::<TokenStream>().unwrap()).unwrap()
+}
+
+/// Derive a label trait
+///
+/// # Args
+///
+/// - `input`: The [`syn::DeriveInput`] for struct that is deriving the label trait
+/// - `trait_path`: The path [`syn::Path`] to the label trait
+pub fn derive_label(input: syn::DeriveInput, trait_path: syn::Path) -> TokenStream {
+    let ident = input.ident;
+
+    let (impl_generics, ty_generics, where_clause) = input.generics.split_for_impl();
+    let mut where_clause = where_clause.cloned().unwrap_or_else(|| syn::WhereClause {
+        where_token: Default::default(),
+        predicates: Default::default(),
+    });
+    where_clause.predicates.push(syn::parse2(quote! { Self: Eq + ::std::fmt::Debug + ::std::hash::Hash + Clone + Send + Sync + 'static }).unwrap());
+
+    (quote! {
+        impl #impl_generics #trait_path for #ident #ty_generics #where_clause {
+            fn dyn_clone(&self) -> Box<dyn #trait_path> {
+                Box::new(Clone::clone(self))
+            }
+        }
+    })
+    .into()
 }
