+++ crates/bevy_utils/src/lib.rs
@@ -1,5 +1,6 @@
 mod enum_variant_meta;
 pub mod slab;
+pub mod label;
 
 pub use ahash::AHasher;
 pub use enum_variant_meta::*;
